+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Breaking changes
 
 - change hasSuperUser to DomainType
+- Open domain's admin is changed from zero address to normal address
 
 ## v0.2.5